大语言模型调用/requirements.txt
rocketmq-client-python
pymysql
requests